Update base_secret_sharing_sum
Split encryption algorithm
